% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{covid}
\alias{covid}
\title{covid}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15861 rows and 5 columns.
}
\usage{
covid
}
\description{
covid
}
\keyword{datasets}
